simplify as possible index cleaners

+++ Read Model/ReadModel.Infrastructure/Indexes/IndexCleaner.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Nest;
 using ReadModel.Domain;
@@ -11,36 +12,30 @@ namespace ReadModel.Infrastructure.Indexes
     public class IndexCleaner : IIndexCleaner
     {
         private readonly ElasticClient _client;
+        private Dictionary<IndexType, Func<Guid?, Task<ResponseBase>>> _indexCleaners;
 
-        public IndexCleaner(IReadModelClient readModelClient) => _client = readModelClient.ElasticClient;
-
-        public async Task CleanIndexAsync(IndexType indexType)
+        public IndexCleaner(IReadModelClient readModelClient)
         {
-            switch (indexType)
+            _client = readModelClient.ElasticClient;
+            _indexCleaners = new Dictionary<IndexType, Func<Guid?, Task<ResponseBase>>>
             {
-                case IndexType.HttpRequest: await CleanIndexAsync<HttpRequest>(); break;
-                case IndexType.Item: await CleanIndexAsync<Item>(); break;
-                default: throw new NotImplementedException();
-            }
+                { IndexType.Item, id => CleanIndexAsync<Item>(id) },
+                { IndexType.HttpRequest, id => CleanIndexAsync<HttpRequest>(id) },
+            };
         }
 
-        public async Task CleanIndexAsync(IndexType indexType, Guid id)
+        public async Task CleanIndexAsync(IndexType indexType) => await _indexCleaners[indexType](null);
+        public async Task CleanIndexAsync(IndexType indexType, Guid id) => await _indexCleaners[indexType](id);
+
+        private async Task<ResponseBase> CleanIndexAsync<TIndex>(Guid? id) where TIndex : class
         {
-            switch (indexType)
-            {
-                case IndexType.HttpRequest: await CleanIndexAsync<HttpRequest>(id); break;
-                case IndexType.Item: await CleanIndexAsync<Item>(id); break;
-                default: throw new NotImplementedException();
-            }
+            if (id is null)
+                return await _client.DeleteByQueryAsync<TIndex>(deleteByQueryDescriptor =>
+                    deleteByQueryDescriptor.Query(queryContainerDescriptor =>
+                        queryContainerDescriptor.QueryString(queryStringQueryDescriptor =>
+                            queryStringQueryDescriptor.Query("*"))));
+            
+            return await _client.DeleteAsync<TIndex>(id);
         }
-
-        private async Task<DeleteByQueryResponse> CleanIndexAsync<TIndex>() where TIndex : class =>
-            await _client.DeleteByQueryAsync<TIndex>(deleteByQueryDescriptor =>
-                deleteByQueryDescriptor.Query(queryContainerDescriptor =>
-                    queryContainerDescriptor.QueryString(queryStringQueryDescriptor =>
-                        queryStringQueryDescriptor.Query("*"))));
-
-        private async Task<DeleteResponse> CleanIndexAsync<TIndex>(Guid id) where TIndex : class =>
-            await _client.DeleteAsync<TIndex>(id);
     }
 }
